Move test to 1.dev
