+++ ledger/ledger-api-tests/suites/src/main/scala/com/daml/ledger/api/testtool/suites/v1_8.scala
@@ -23,7 +23,6 @@ package object v1_8 {
       new ContractKeysIT,
       new DeeplyNestedValueIT,
       new DivulgenceIT,
-      new ExplicitDisclosureIT,
       new HealthServiceIT,
       new IdentityIT,
       new LedgerConfigurationServiceIT,
