+++ ledger/ledger-api-tests/suites/src/main/scala/com/daml/ledger/api/testtool/suites/v1_dev.scala
@@ -8,7 +8,10 @@ import com.daml.ledger.api.tls.TlsConfiguration
 
 package object v1_dev {
   def default(timeoutScaleFactor: Double): Vector[LedgerTestSuite] =
-    v1_14.default(timeoutScaleFactor) :+ new InterfaceIT
+    v1_14.default(timeoutScaleFactor) ++ Vector(
+      new InterfaceIT,
+      new ExplicitDisclosureIT,
+    )
 
   def optional(tlsConfig: Option[TlsConfiguration]): Vector[LedgerTestSuite] =
     v1_14.optional(tlsConfig)
