+++ ledger/sandbox-on-x/BUILD.bazel
@@ -534,3 +534,21 @@ conformance_test(
         "--verbose",
     ],
 )
+
+# TODO DPP-1026: Make sure ExplicitDisclosureIT is run on all servers (it's guarded by a feature flag)
+server_conformance_test(
+    name = "conformance-test-explicit-disclosure",
+    lf_versions = [
+        "1.dev",
+    ],
+    server_args = [
+        "--contract-id-seeding=testing-weak",
+        "--participant participant-id=example,port=6865",
+        "--enable-user-management=true",
+    ],
+    servers = SERVERS,
+    test_tool_args = [
+        "--verbose",
+        "--include=ExplicitDisclosureIT",
+    ],
+)
